Refactor API signatures and minor fixes
Clean up function signatures, remove unnecessary moves/forwards, and fix small portability issues:

- Font: stop moving the DescriptorSetLayout when creating the descriptor set (use the provided layout value).
- PhysicalDevice: mark several query methods [[nodiscard]] to catch ignored return values (isDeviceSuitable, findQueueFamilies, querySwapChainSupport).
- RayTracingPipeline: use auto for stageIndex and remove unused forward declaration of DescriptorSet in the header.
- Renderer2D: default-initialize m_currentFontName instead of using an explicit empty string.
- MousePicker: make getObjectIDFromBuffer non-const in the implementation and declare it static in the header; remove unused forward declarations.
- Window: cast width/height to int when calling glfwCreateWindow to ensure correct overload selection.

These changes improve code clarity, prevent unnecessary moves, reduce warnings, and address minor portability/typing issues.

+++ source/components/window/Window.cpp
@@ -23,7 +23,13 @@ namespace vke {
     }
     else
     {
-      m_window = glfwCreateWindow(config.width, config.height, config.title.c_str(), nullptr, nullptr);
+      m_window = glfwCreateWindow(
+        static_cast<int>(config.width),
+        static_cast<int>(config.height),
+        config.title.c_str(),
+        nullptr,
+        nullptr
+      );
     }
 
     if (m_window == nullptr)
